+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Patch Changes
 
-- Updated dependencies
+- Updated dependencies [d5cca1e8]
+- Updated dependencies [f4b5c0df]
   - @ballerine/common@0.7.24
 
 ## 1.4.20